requirements.txt
djangorestframework
django-rest-auth
pillow
django-cors-headers
